requirements/test.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    make upgrade
#
aiohappyeyeballs==2.4.6
    # via
    #   -r requirements/base.txt
    #   aiohttp
aiohttp==3.11.13
    # via
    #   -r requirements/base.txt
    #   geoip2
aiosignal==1.3.2
    # via
    #   -r requirements/base.txt
    #   aiohttp
asgiref==3.8.1
    # via
    #   -r requirements/base.txt
    #   django
attrs==25.1.0
    # via
    #   -r requirements/base.txt
    #   aiohttp
certifi==2025.1.31
    # via
    #   -r requirements/base.txt
    #   requests
charset-normalizer==3.4.1
    # via
    #   -r requirements/base.txt
    #   requests
click==8.1.8
    # via code-annotations
code-annotations==2.2.0
    # via -r requirements/test.in
coverage[toml]==7.6.12
    # via pytest-cov
ddt==1.7.2
    # via -r requirements/test.in
    # via
    #   -c https://raw.githubusercontent.com/edx/edx-lint/master/edx_lint/files/common_constraints.txt
    #   -r requirements/base.txt
django-ipware==7.0.1
    # via -r requirements/base.txt
factory-boy==3.3.3
    # via -r requirements/test.in
faker==36.1.1
    # via factory-boy
frozenlist==1.5.0
    # via
    #   -r requirements/base.txt
    #   aiohttp
    #   aiosignal
geoip2==5.0.1
    # via -r requirements/base.txt
idna==3.10
    # via
    #   -r requirements/base.txt
    #   requests
    #   yarl
iniconfig==2.0.0
    # via pytest
jinja2==3.1.5
    # via code-annotations
markupsafe==3.0.2
    # via jinja2
maxminddb==2.6.3
    # via
    #   -r requirements/base.txt
    #   geoip2
multidict==6.1.0
    # via
    #   -r requirements/base.txt
    #   aiohttp
    #   yarl
packaging==24.2
    # via pytest
path==17.1.0
    # via -r requirements/base.txt
pbr==6.1.1
    # via stevedore
pluggy==1.5.0
    # via pytest
propcache==0.3.0
    # via
    #   -r requirements/base.txt
    #   aiohttp
    #   yarl
pytest==8.3.5
    # via
    #   pytest-cov
    #   pytest-django
pytest-cov==6.0.0
    # via -r requirements/test.in
pytest-django==4.10.0
    # via -r requirements/test.in
python-ipware==3.0.0
    # via
    #   -r requirements/base.txt
    #   django-ipware
python-slugify==8.0.4
    # via code-annotations
pyyaml==6.0.2
    # via code-annotations
requests==2.32.3
    # via
    #   -r requirements/base.txt
    #   geoip2
sqlparse==0.5.3
    # via
    #   -r requirements/base.txt
    #   django
stevedore==5.4.1
    # via code-annotations
text-unidecode==1.3
    # via python-slugify
tzdata==2025.1
    # via faker
urllib3==2.2.3
    # via
    #   -c https://raw.githubusercontent.com/edx/edx-lint/master/edx_lint/files/common_constraints.txt
    #   -r requirements/base.txt
    #   requests
yarl==1.18.3
    # via
    #   -r requirements/base.txt
    #   aiohttp

# The following packages are considered to be unsafe in a requirements file:
# setuptools
